fix: some bug

+++ src/main/java/cn/nukkit/network/protocol/UpdateBlockPacket.java
@@ -9,20 +9,18 @@ import lombok.ToString;
 @ToString
 public class UpdateBlockPacket extends DataPacket {
     public static final int NETWORK_ID = ProtocolInfo.UPDATE_BLOCK_PACKET;
-
     public static final int FLAG_NONE = 0b0000;
     public static final int FLAG_NEIGHBORS = 0b0001;
     public static final int FLAG_NETWORK = 0b0010;
     public static final int FLAG_NOGRAPHIC = 0b0100;
     public static final int FLAG_PRIORITY = 0b1000;
-
     public static final int FLAG_ALL = (FLAG_NEIGHBORS | FLAG_NETWORK);
     public static final int FLAG_ALL_PRIORITY = (FLAG_ALL | FLAG_PRIORITY);
 
     public int x;
     public int z;
     public int y;
-    public int blockRuntimeId;
+    public long blockRuntimeId;
     public int flags;
     public int dataLayer = 0;
